+++ types/types.d.ts
@@ -43,3 +43,8 @@ type VersionDB = {
   id: string;
   name: Version;
 };
+
+type Catch = {
+  name: string;
+  uri: string;
+};
